+++ test/compilationTests/gnosis/Utils/Math.sol
@@ -10,9 +10,9 @@ library Math {
      *  Constants
      */
     // This is equal to 1 in our calculations
-    uint public constant ONE =  0x10000000000000000;
-    uint public constant LN2 = 0xb17217f7d1cf79ac;
-    uint public constant LOG2_E = 0x171547652b82fe177;
+    uint256 public constant ONE =  0x10000000000000000;
+    uint256 public constant LN2 = 0xb17217f7d1cf79ac;
+    uint256 public constant LOG2_E = 0x171547652b82fe177;
 
     /*
      *  Public functions
@@ -20,32 +20,32 @@ library Math {
     /// @dev Returns natural exponential function value of given x
     /// @param x x
     /// @return e**x
-    function exp(int x)
+    function exp(int256 x)
         public
         constant
-        returns (uint)
+        returns (uint256)
     {
         // revert if x is > MAX_POWER, where
-        // MAX_POWER = int(mp.floor(mp.log(mpf(2**256 - 1) / ONE) * ONE))
+        // MAX_POWER = int256(mp.floor(mp.log(mpf(2**256 - 1) / ONE) * ONE))
         require(x <= 2454971259878909886679);
         // return 0 if exp(x) is tiny, using
-        // MIN_POWER = int(mp.floor(mp.log(mpf(1) / ONE) * ONE))
+        // MIN_POWER = int256(mp.floor(mp.log(mpf(1) / ONE) * ONE))
         if (x < -818323753292969962227)
             return 0;
         // Transform so that e^x -> 2^x
-        x = x * int(ONE) / int(LN2);
+        x = x * int256(ONE) / int256(LN2);
         // 2^x = 2^whole(x) * 2^frac(x)
         //       ^^^^^^^^^^ is a bit shift
         // so Taylor expand on z = frac(x)
-        int shift;
-        uint z;
+        int256 shift;
+        uint256 z;
         if (x >= 0) {
-            shift = x / int(ONE);
-            z = uint(x % int(ONE));
+            shift = x / int256(ONE);
+            z = uint256(x % int256(ONE));
         }
         else {
-            shift = x / int(ONE) - 1;
-            z = ONE - uint(-x % int(ONE));
+            shift = x / int256(ONE) - 1;
+            z = ONE - uint256(-x % int256(ONE));
         }
         // 2^x = 1 + (ln 2) x + (ln 2)^2/2! x^2 + ...
         //
@@ -53,15 +53,15 @@ library Math {
         // >>> from mpmath import mp
         // >>> mp.dps = 100
         // >>> ONE =  0x10000000000000000
-        // >>> print('\n'.join(hex(int(mp.log(2)**i / mp.factorial(i) * ONE)) for i in range(1, 7)))
+        // >>> print('\n'.join(hex(int256(mp.log(2)**i / mp.factorial(i) * ONE)) for i in range(1, 7)))
         // 0xb17217f7d1cf79ab
         // 0x3d7f7bff058b1d50
         // 0xe35846b82505fc5
         // 0x276556df749cee5
         // 0x5761ff9e299cc4
         // 0xa184897c363c3
-        uint zpow = z;
-        uint result = ONE;
+        uint256 zpow = z;
+        uint256 result = ONE;
         result += 0xb17217f7d1cf79ab * zpow / ONE;
         zpow = zpow * z / ONE;
         result += 0x3d7f7bff058b1d50 * zpow / ONE;
@@ -105,67 +105,67 @@ library Math {
     /// @dev Returns natural logarithm value of given x
     /// @param x x
     /// @return ln(x)
-    function ln(uint x)
+    function ln(uint256 x)
         public
         constant
-        returns (int)
+        returns (int256)
     {
         require(x > 0);
         // binary search for floor(log2(x))
-        int ilog2 = floorLog2(x);
-        int z;
+        int256 ilog2 = floorLog2(x);
+        int256 z;
         if (ilog2 < 0)
-            z = int(x << uint(-ilog2));
+            z = int256(x << uint256(-ilog2));
         else
-            z = int(x >> uint(ilog2));
+            z = int256(x >> uint256(ilog2));
         // z = x * 2^-⌊log₂x⌋
         // so 1 <= z < 2
         // and ln z = ln x - ⌊log₂x⌋/log₂e
         // so just compute ln z using artanh series
         // and calculate ln x from that
-        int term = (z - int(ONE)) * int(ONE) / (z + int(ONE));
-        int halflnz = term;
-        int termpow = term * term / int(ONE) * term / int(ONE);
+        int256 term = (z - int256(ONE)) * int256(ONE) / (z + int256(ONE));
+        int256 halflnz = term;
+        int256 termpow = term * term / int256(ONE) * term / int256(ONE);
         halflnz += termpow / 3;
-        termpow = termpow * term / int(ONE) * term / int(ONE);
+        termpow = termpow * term / int256(ONE) * term / int256(ONE);
         halflnz += termpow / 5;
-        termpow = termpow * term / int(ONE) * term / int(ONE);
+        termpow = termpow * term / int256(ONE) * term / int256(ONE);
         halflnz += termpow / 7;
-        termpow = termpow * term / int(ONE) * term / int(ONE);
+        termpow = termpow * term / int256(ONE) * term / int256(ONE);
         halflnz += termpow / 9;
-        termpow = termpow * term / int(ONE) * term / int(ONE);
+        termpow = termpow * term / int256(ONE) * term / int256(ONE);
         halflnz += termpow / 11;
-        termpow = termpow * term / int(ONE) * term / int(ONE);
+        termpow = termpow * term / int256(ONE) * term / int256(ONE);
         halflnz += termpow / 13;
-        termpow = termpow * term / int(ONE) * term / int(ONE);
+        termpow = termpow * term / int256(ONE) * term / int256(ONE);
         halflnz += termpow / 15;
-        termpow = termpow * term / int(ONE) * term / int(ONE);
+        termpow = termpow * term / int256(ONE) * term / int256(ONE);
         halflnz += termpow / 17;
-        termpow = termpow * term / int(ONE) * term / int(ONE);
+        termpow = termpow * term / int256(ONE) * term / int256(ONE);
         halflnz += termpow / 19;
-        termpow = termpow * term / int(ONE) * term / int(ONE);
+        termpow = termpow * term / int256(ONE) * term / int256(ONE);
         halflnz += termpow / 21;
-        termpow = termpow * term / int(ONE) * term / int(ONE);
+        termpow = termpow * term / int256(ONE) * term / int256(ONE);
         halflnz += termpow / 23;
-        termpow = termpow * term / int(ONE) * term / int(ONE);
+        termpow = termpow * term / int256(ONE) * term / int256(ONE);
         halflnz += termpow / 25;
-        return (ilog2 * int(ONE)) * int(ONE) / int(LOG2_E) + 2 * halflnz;
+        return (ilog2 * int256(ONE)) * int256(ONE) / int256(LOG2_E) + 2 * halflnz;
     }
 
     /// @dev Returns base 2 logarithm value of given x
     /// @param x x
     /// @return logarithmic value
-    function floorLog2(uint x)
+    function floorLog2(uint256 x)
         public
         constant
-        returns (int lo)
+        returns (int256 lo)
     {
         lo = -64;
-        int hi = 193;
+        int256 hi = 193;
         // I use a shift here instead of / 2 because it floors instead of rounding towards 0
-        int mid = (hi + lo) >> 1;
+        int256 mid = (hi + lo) >> 1;
         while((lo + 1) < hi) {
-            if (mid < 0 && x << uint(-mid) < ONE || mid >= 0 && x >> uint(mid) < ONE)
+            if (mid < 0 && x << uint256(-mid) < ONE || mid >= 0 && x >> uint256(mid) < ONE)
                 hi = mid;
             else
                 lo = mid;
@@ -176,14 +176,14 @@ library Math {
     /// @dev Returns maximum of an array
     /// @param nums Numbers to look through
     /// @return Maximum number
-    function max(int[] nums)
+    function max(int256[] nums)
         public
         constant
-        returns (int max)
+        returns (int256 max)
     {
         require(nums.length > 0);
         max = -2**255;
-        for (uint i = 0; i < nums.length; i++)
+        for (uint256 i = 0; i < nums.length; i++)
             if (nums[i] > max)
                 max = nums[i];
     }
@@ -192,7 +192,7 @@ library Math {
     /// @param a First addend
     /// @param b Second addend
     /// @return Did no overflow occur?
-    function safeToAdd(uint a, uint b)
+    function safeToAdd(uint256 a, uint256 b)
         public
         constant
         returns (bool)
@@ -204,7 +204,7 @@ library Math {
     /// @param a Minuend
     /// @param b Subtrahend
     /// @return Did no underflow occur?
-    function safeToSub(uint a, uint b)
+    function safeToSub(uint256 a, uint256 b)
         public
         constant
         returns (bool)
@@ -216,7 +216,7 @@ library Math {
     /// @param a First factor
     /// @param b Second factor
     /// @return Did no overflow occur?
-    function safeToMul(uint a, uint b)
+    function safeToMul(uint256 a, uint256 b)
         public
         constant
         returns (bool)
@@ -228,10 +228,10 @@ library Math {
     /// @param a First addend
     /// @param b Second addend
     /// @return Sum
-    function add(uint a, uint b)
+    function add(uint256 a, uint256 b)
         public
         constant
-        returns (uint)
+        returns (uint256)
     {
         require(safeToAdd(a, b));
         return a + b;
@@ -241,10 +241,10 @@ library Math {
     /// @param a Minuend
     /// @param b Subtrahend
     /// @return Difference
-    function sub(uint a, uint b)
+    function sub(uint256 a, uint256 b)
         public
         constant
-        returns (uint)
+        returns (uint256)
     {
         require(safeToSub(a, b));
         return a - b;
@@ -254,10 +254,10 @@ library Math {
     /// @param a First factor
     /// @param b Second factor
     /// @return Product
-    function mul(uint a, uint b)
+    function mul(uint256 a, uint256 b)
         public
         constant
-        returns (uint)
+        returns (uint256)
     {
         require(safeToMul(a, b));
         return a * b;
@@ -267,7 +267,7 @@ library Math {
     /// @param a First addend
     /// @param b Second addend
     /// @return Did no overflow occur?
-    function safeToAdd(int a, int b)
+    function safeToAdd(int256 a, int256 b)
         public
         constant
         returns (bool)
@@ -279,7 +279,7 @@ library Math {
     /// @param a Minuend
     /// @param b Subtrahend
     /// @return Did no underflow occur?
-    function safeToSub(int a, int b)
+    function safeToSub(int256 a, int256 b)
         public
         constant
         returns (bool)
@@ -291,7 +291,7 @@ library Math {
     /// @param a First factor
     /// @param b Second factor
     /// @return Did no overflow occur?
-    function safeToMul(int a, int b)
+    function safeToMul(int256 a, int256 b)
         public
         constant
         returns (bool)
@@ -303,10 +303,10 @@ library Math {
     /// @param a First addend
     /// @param b Second addend
     /// @return Sum
-    function add(int a, int b)
+    function add(int256 a, int256 b)
         public
         constant
-        returns (int)
+        returns (int256)
     {
         require(safeToAdd(a, b));
         return a + b;
@@ -316,10 +316,10 @@ library Math {
     /// @param a Minuend
     /// @param b Subtrahend
     /// @return Difference
-    function sub(int a, int b)
+    function sub(int256 a, int256 b)
         public
         constant
-        returns (int)
+        returns (int256)
     {
         require(safeToSub(a, b));
         return a - b;
@@ -329,10 +329,10 @@ library Math {
     /// @param a First factor
     /// @param b Second factor
     /// @return Product
-    function mul(int a, int b)
+    function mul(int256 a, int256 b)
         public
         constant
-        returns (int)
+        returns (int256)
     {
         require(safeToMul(a, b));
         return a * b;
